ai/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
asyncio-mqtt==0.16.2
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
billiard==4.2.3
black==25.11.0
bleach==6.3.0
blinker==1.9.0
branca==0.8.2
cachetools==6.2.2
catboost==1.2.8
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmdstanpy==1.3.0
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
efficientnet==1.1.1
efficientnet_pytorch==0.7.1
executing==2.2.1
fastapi==0.121.3
fastjsonschema==2.21.2
filelock==3.20.0
findspark==2.0.1
flake8==7.3.0
Flask==3.0.0
flask-cors==6.0.1
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
h5py==3.15.1
holidays==0.85
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageHash==4.3.2
ImageIO==2.37.2
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keras==3.12.0
Keras-Applications==1.0.8
kiwisolver==1.4.9
kombu==5.5.4
lark==1.3.1
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
loguru==0.7.3
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.12.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.5.0
notebook_shim==0.2.4
numpy==1.26.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.18.0
optuna==4.6.0
overrides==7.7.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
peewee==3.18.3
pickle-mixin==1.0.2
Pillow==10.1.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
prophet==1.2.1
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.1.3
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pymongo==4.15.4
pyparsing==3.2.5
pyspark==3.5.7
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyWavelets==1.9.0
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
reportlab==4.0.7
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.29.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.45.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
stanio==0.5.1
starlette==0.50.0
stringzilla==4.2.3
sympy==1.14.0
ta==0.11.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.15
win32_setctime==1.2.0
wrapt==2.0.1
xgboost==3.1.1
xyzservices==2025.10.0
yarl==1.22.0
yfinance==0.2.66
"Flask-Cors" 
